+++ mesh_protocol_dwm1001-dev/src/LCG.c
@@ -45,7 +45,7 @@ LCG LCG_Create(uint32_t seed) {
   return self;
 };
 
-int LCG_Rand(Node node) {
+uint8_t LCG_Rand(Node node) {
   /** reimplementation of rand from https://stackoverflow.com/a/10198842 by Matteo Italia */
   node->lcg->next = node->lcg->next * 1103515245 + 12345;
   return (uint32_t)((node->lcg->next/65536) % 32768);
